@@ -6,7 +6,7 @@ import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
 class CalculatorScreen extends StatelessWidget {
-  const CalculatorScreen({Key key}) : super(key: key);
+  const CalculatorScreen() : super();
 
   @override
   Widget build(BuildContext context) {
