@@ -4,7 +4,7 @@ import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
 class Results extends StatelessWidget {
-  const Results({Key key}) : super(key: key);
+  const Results() : super();
 
   @override
   Widget build(BuildContext context) {
